Move database datatype fallbacks and adapters; not exclusive to SQL (#72)
