+++ core/src/main/java/org/minerift/ether/database/DataTypeFallbacks.java
@@ -1,6 +1,6 @@
 package org.minerift.ether.database;
 
-import org.minerift.ether.database.sql.fallback.Fallback;
+import org.minerift.ether.database.fallback.Fallback;
 
 public interface DataTypeFallbacks {
     <T> Fallback<T, ?> getPossibleFallback(DataType<T> type, DatabaseCreationContext ctx);
