+++ core/src/main/java/org/minerift/ether/database/Field.java
@@ -1,7 +1,7 @@
 package org.minerift.ether.database;
 
-import org.minerift.ether.database.sql.fallback.Fallback;
-import org.minerift.ether.database.sql.adapters.Adapter;
+import org.minerift.ether.database.fallback.Fallback;
+import org.minerift.ether.database.adapters.Adapter;
 import org.minerift.ether.util.Utils;
 
 import java.lang.ref.WeakReference;
