+++ core/src/main/java/org/minerift/ether/database/adapters/EtherUsers2UuidsAdapter.java
@@ -1,4 +1,4 @@
-package org.minerift.ether.database.sql.adapters;
+package org.minerift.ether.database.adapters;
 
 import org.minerift.ether.user.EtherUser;
