+++ core/src/main/java/org/minerift/ether/database/fallback/Fallback.java
@@ -1,7 +1,7 @@
-package org.minerift.ether.database.sql.fallback;
+package org.minerift.ether.database.fallback;
 
 import org.minerift.ether.database.DataType;
-import org.minerift.ether.database.sql.adapters.Adapter;
+import org.minerift.ether.database.adapters.Adapter;
 
 public abstract class Fallback<T, F> implements Adapter<T, F> {
