+++ core/src/main/java/org/minerift/ether/database/sql/SqlDataTypeFallbacks.java
@@ -3,7 +3,7 @@ package org.minerift.ether.database.sql;
 import org.minerift.ether.database.DataType;
 import org.minerift.ether.database.DataTypeFallbacks;
 import org.minerift.ether.database.DatabaseCreationContext;
-import org.minerift.ether.database.sql.fallback.Fallback;
+import org.minerift.ether.database.fallback.Fallback;
 
 public class SqlDataTypeFallbacks implements DataTypeFallbacks {
     @Override
